将价格类型转换为double，完成了财务相关功能

+++ src/main/Java/com/netease/sale/dao/BuyDao.java
@@ -21,7 +21,7 @@ public interface BuyDao {
      * @param buyTime
      * @return
      */
-    public void addBuy(@Param("buyerId") int buyerId, @Param("goodsId") int goodsId, @Param("number") int number, @Param("buyTime") String buyTime, @Param("oldPrice")BigDecimal oldPrice);
+    public void addBuy(@Param("buyerId") int buyerId, @Param("goodsId") int goodsId, @Param("number") int number, @Param("buyTime") String buyTime, @Param("oldPrice") Double oldPrice);
 
     /**
      * 查询用户购买记录简要信息
